requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-django>=4.5.0
flake8>=5.0.0
bandit>=1.7.0
safety>=2.0.0
coverage>=6.0.0
